Basic support for gopherjs/wasm and webgl backend.
To try, you can use gopherjs build for one of the command, but the nicest way
is with `gopherjs serve fyne.io/fyne` that will open a local web server that you
can use to navigate fyne directory and check any app with it. Otherwise you can
build with GOOS=js GOARCH=wasm go build for the WebAssembly backend.
