+++ app/app_other.go
@@ -1,4 +1,4 @@
-// +build ci !linux,!darwin,!windows,!freebsd,!openbsd,!netbsd
+// +build ci !linux,!darwin,!windows,!freebsd,!openbsd,!netbsd,!js,!wasm,!web
 
 package app
