+++ app/settings.go
@@ -1,7 +1,6 @@
 package app
 
 import (
-	"encoding/json"
 	"os"
 	"path/filepath"
 	"sync"
@@ -81,33 +80,6 @@ func (s *settings) apply() {
 	}
 }
 
-func (s *settings) load() {
-	err := s.loadFromFile(s.schema.StoragePath())
-	if err != nil {
-		fyne.LogError("Settings load error:", err)
-	}
-
-	s.setupTheme()
-}
-
-func (s *settings) loadFromFile(path string) error {
-	file, err := os.Open(path) // #nosec
-	if err != nil {
-		if os.IsNotExist(err) {
-			return nil
-		}
-		return err
-	}
-	decode := json.NewDecoder(file)
-
-	return decode.Decode(&s.schema)
-}
-
-func (s *settings) fileChanged() {
-	s.load()
-	s.apply()
-}
-
 func (s *settings) setupTheme() {
 	if s.themeSpecified {
 		return
@@ -132,4 +104,9 @@ func loadSettings() *settings {
 	s.load()
 
 	return s
+}
+
+func (s *settings) fileChanged() {
+	s.load()
+	s.apply()
 }
