+++ internal/driver/glfw/glfw_webgl.go
@@ -1,0 +1,6 @@
+// +build js wasm web
+
+package glfw
+
+func initWindowHints() {
+}
