+++ internal/painter/gl/gl_goxjs.go
@@ -1,5 +1,4 @@
-// +build gles arm arm64
-// +build !android,!ios,!mobile,!js,!wasm,!web
+// +build js wasm web
 
 package gl
 
@@ -7,30 +6,30 @@ import (
 	"fmt"
 	"image"
 	"image/draw"
-	"strings"
 
-	gl "github.com/go-gl/gl/v3.1/gles2"
+	"encoding/binary"
+	"golang.org/x/mobile/exp/f32"
+	"github.com/goxjs/gl"
 
 	"fyne.io/fyne"
 	"fyne.io/fyne/theme"
 )
 
 // Buffer represents a GL buffer
-type Buffer uint32
+type Buffer gl.Buffer
 
 // Program represents a compiled GL program
-type Program uint32
+type Program gl.Program
 
 // Texture represents an uploaded GL texture
-type Texture uint32
+type Texture gl.Texture
 
 // NoTexture is the zero value for a Texture
-var NoTexture = Texture(0)
+var NoTexture = Texture(gl.NoTexture)
 
-func newTexture() Texture {
-	var texture uint32
+func (p *glPainter) newTexture() Texture {
+	var texture = gl.CreateTexture()
 
-	gl.GenTextures(1, &texture)
 	gl.ActiveTexture(gl.TEXTURE0)
 	gl.BindTexture(gl.TEXTURE_2D, texture)
 	gl.TexParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR)
@@ -48,7 +47,7 @@ func getTexture(object fyne.CanvasObject, creator func(canvasObject fyne.CanvasO
 		texture = creator(object)
 		textures[object] = texture
 	}
-	if texture == NoTexture {
+	if !gl.Texture(texture).Valid() {
 		return NoTexture, fmt.Errorf("No texture available.")
 	}
 	return texture, nil
@@ -57,21 +56,19 @@ func getTexture(object fyne.CanvasObject, creator func(canvasObject fyne.CanvasO
 func (p *glPainter) imgToTexture(img image.Image) Texture {
 	switch i := img.(type) {
 	case *image.Uniform:
-		texture := newTexture()
+		texture := p.newTexture()
 		r, g, b, a := i.RGBA()
 		r8, g8, b8, a8 := uint8(r>>8), uint8(g>>8), uint8(b>>8), uint8(a>>8)
 		data := []uint8{r8, g8, b8, a8}
-		gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGBA, 1, 1, 0, gl.RGBA,
-			gl.UNSIGNED_BYTE, gl.Ptr(data))
+		gl.TexImage2D(gl.TEXTURE_2D, 0, 1, 1, gl.RGBA, gl.UNSIGNED_BYTE, data)
 		return texture
 	case *image.RGBA:
 		if len(i.Pix) == 0 { // image is empty
-			return 0
+			return NoTexture
 		}
-
-		texture := newTexture()
-		gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGBA, int32(i.Rect.Size().X), int32(i.Rect.Size().Y),
-			0, gl.RGBA, gl.UNSIGNED_BYTE, gl.Ptr(i.Pix))
+		texture := p.newTexture()
+		gl.TexImage2D(gl.TEXTURE_2D, 0, i.Rect.Size().X, i.Rect.Size().Y,
+			gl.RGBA, gl.UNSIGNED_BYTE, i.Pix)
 		return texture
 	default:
 		rgba := image.NewRGBA(image.Rect(0, 0, img.Bounds().Dx(), img.Bounds().Dy()))
@@ -81,47 +78,28 @@ func (p *glPainter) imgToTexture(img image.Image) Texture {
 }
 
 func (p *glPainter) SetOutputSize(width, height int) {
-	gl.Viewport(0, 0, int32(width), int32(height))
+	gl.Viewport(0, 0, width, height)
 }
 
 func (p *glPainter) freeTexture(obj fyne.CanvasObject) {
-	texture := textures[obj]
-	if texture != 0 {
-		tex := uint32(texture)
-		gl.DeleteTextures(1, &tex)
+	texture, ok := textures[obj]
+	if ok {
+		gl.DeleteTexture(gl.Texture(texture))
 		delete(textures, obj)
 	}
 }
 
-func glInit() {
-	err := gl.Init()
-	if err != nil {
-		fyne.LogError("failed to initialise OpenGL", err)
-		return
-	}
-
-	gl.Disable(gl.DEPTH_TEST)
-	gl.Enable(gl.BLEND)
-}
-
-func compileShader(source string, shaderType uint32) (uint32, error) {
+func (p *glPainter) compileShader(source string, shaderType gl.Enum) (gl.Shader, error) {
 	shader := gl.CreateShader(shaderType)
 
-	csources, free := gl.Strs(source)
-	gl.ShaderSource(shader, 1, csources, nil)
-	free()
+	gl.ShaderSource(shader, source)
 	gl.CompileShader(shader)
 
-	var status int32
-	gl.GetShaderiv(shader, gl.COMPILE_STATUS, &status)
+	status := gl.GetShaderi(shader, gl.COMPILE_STATUS)
 	if status == gl.FALSE {
-		var logLength int32
-		gl.GetShaderiv(shader, gl.INFO_LOG_LENGTH, &logLength)
+		info := gl.GetShaderInfoLog(shader)
 
-		info := strings.Repeat("\x00", int(logLength+1))
-		gl.GetShaderInfoLog(shader, logLength, nil, gl.Str(info))
-
-		return 0, fmt.Errorf("failed to compile %v: %v", source, info)
+		return shader, fmt.Errorf("failed to compile %v: %v", source, info)
 	}
 
 	return shader, nil
@@ -138,8 +116,7 @@ const (
         fragTexCoord = vertTexCoord;
 
         gl_Position = vec4(vert, 1);
-    }
-` + "\x00"
+    }`
 
 	fragmentShaderSource = `
     #version 100
@@ -149,16 +126,17 @@ const (
 
     void main() {
         gl_FragColor = texture2D(tex, fragTexCoord);
-    }
-` + "\x00"
+    }`
 )
 
 func (p *glPainter) Init() {
-	vertexShader, err := compileShader(vertexShaderSource, gl.VERTEX_SHADER)
+	gl.Disable(gl.DEPTH_TEST)
+
+	vertexShader, err := p.compileShader(vertexShaderSource, gl.VERTEX_SHADER)
 	if err != nil {
 		panic(err)
 	}
-	fragmentShader, err := compileShader(fragmentShaderSource, gl.FRAGMENT_SHADER)
+	fragmentShader, err := p.compileShader(fragmentShaderSource, gl.FRAGMENT_SHADER)
 	if err != nil {
 		panic(err)
 	}
@@ -172,8 +150,6 @@ func (p *glPainter) Init() {
 }
 
 func (p *glPainter) glClearBuffer() {
-	gl.UseProgram(uint32(p.program))
-
 	r, g, b, a := theme.BackgroundColor().RGBA()
 	max16bit := float32(255 * 255)
 	gl.ClearColor(float32(r)/max16bit, float32(g)/max16bit, float32(b)/max16bit, float32(a)/max16bit)
@@ -190,29 +166,30 @@ func (p *glPainter) glScissorClose() {
 }
 
 func (p *glPainter) glCreateBuffer(points []float32) Buffer {
-	var vbo uint32
-	gl.GenBuffers(1, &vbo)
+	vbo := gl.CreateBuffer()
 	gl.BindBuffer(gl.ARRAY_BUFFER, vbo)
-	gl.BufferData(gl.ARRAY_BUFFER, 4*len(points), gl.Ptr(points), gl.STATIC_DRAW)
+	gl.BufferData(gl.ARRAY_BUFFER, f32.Bytes(binary.LittleEndian, points...), gl.STATIC_DRAW)
 
-	vertAttrib := uint32(gl.GetAttribLocation(uint32(p.program), gl.Str("vert\x00")))
+	vertAttrib := gl.GetAttribLocation(gl.Program(p.program), "vert")
 	gl.EnableVertexAttribArray(vertAttrib)
-	gl.VertexAttribPointer(vertAttrib, 3, gl.FLOAT, false, 5*4, gl.PtrOffset(0))
+	gl.VertexAttribPointer(vertAttrib, 3, gl.FLOAT, false, 5*4, 0)
 
-	texCoordAttrib := uint32(gl.GetAttribLocation(uint32(p.program), gl.Str("vertTexCoord\x00")))
+	texCoordAttrib := gl.GetAttribLocation(gl.Program(p.program), "vertTexCoord")
 	gl.EnableVertexAttribArray(texCoordAttrib)
-	gl.VertexAttribPointer(texCoordAttrib, 2, gl.FLOAT, false, 5*4, gl.PtrOffset(3*4))
+	gl.VertexAttribPointer(texCoordAttrib, 2, gl.FLOAT, false, 5*4, 3*4)
 
 	return Buffer(vbo)
 }
 
 func (p *glPainter) glFreeBuffer(vbo Buffer) {
-	gl.BindBuffer(gl.ARRAY_BUFFER, 0)
-	buf := uint32(vbo)
-	gl.DeleteBuffers(1, &buf)
+	gl.BindBuffer(gl.ARRAY_BUFFER, gl.Buffer(vbo))
+	gl.DeleteBuffer(gl.Buffer(vbo))
 }
 
 func (p *glPainter) glDrawTexture(texture Texture, alpha float32) {
+	gl.UseProgram(gl.Program(p.program))
+	gl.Enable(gl.BLEND)
+
 	// here we have to choose between blending the image alpha or fading it...
 	// TODO find a way to support both
 	if alpha != 1.0 {
@@ -223,12 +200,15 @@ func (p *glPainter) glDrawTexture(texture Texture, alpha float32) {
 	}
 
 	gl.ActiveTexture(gl.TEXTURE0)
-	gl.BindTexture(gl.TEXTURE_2D, uint32(texture))
+	gl.BindTexture(gl.TEXTURE_2D, gl.Texture(texture))
 
 	gl.DrawArrays(gl.TRIANGLE_STRIP, 0, 4)
 }
 
 func (p *glPainter) glCapture(width, height int32, pixels *[]uint8) {
-	gl.ReadBuffer(gl.FRONT)
-	gl.ReadPixels(0, 0, int32(width), int32(height), gl.RGBA, gl.UNSIGNED_BYTE, gl.Ptr(*pixels))
+	gl.ReadPixels(*pixels, 0, 0, int(width), int(height), gl.RGBA, gl.UNSIGNED_BYTE)
+}
+
+func glInit() {
+	// no-op, gomobile does this
 }
